feat: Add script to retrieve Star Wars movie title
The code changes add a new JavaScript file `3-starwars_title.js` that retrieves the title of a Star Wars movie using the SWAPI API. The script takes an episode ID as a command-line argument and makes a request to the API to fetch the movie details. It then parses the response and logs the movie title to the console.

Based on the recent user commits and repository commits, it seems that the repository follows a convention of starting commit messages with a verb in the imperative form. Therefore, the suggested commit message is "Add script to retrieve Star Wars movie title".